+++ src/libromdata/Console/saturn_structs.h
@@ -1,8 +1,8 @@
 /***************************************************************************
  * ROM Properties Page shell extension. (libromdata)                       *
- * dc_structs.h: Sega Dreamcast data structures.                           *
+ * saturn_structs.h: Sega Saturn data structures.                          *
  *                                                                         *
- * Copyright (c) 2017 by David Korth.                                      *
+ * Copyright (c) 2017-2018 by David Korth.                                 *
  *                                                                         *
  * This program is free software; you can redistribute it and/or modify it *
  * under the terms of the GNU General Public License as published by the   *
@@ -14,13 +14,12 @@
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the           *
  * GNU General Public License for more details.                            *
  *                                                                         *
- * You should have received a copy of the GNU General Public License along *
- * with this program; if not, write to the Free Software Foundation, Inc., *
- * 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.           *
+ * You should have received a copy of the GNU General Public License       *
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.   *
  ***************************************************************************/
 
-#ifndef __ROMPROPERTIES_LIBROMDATA_DC_STRUCTS_H__
-#define __ROMPROPERTIES_LIBROMDATA_DC_STRUCTS_H__
+#ifndef __ROMPROPERTIES_LIBROMDATA_SATURN_STRUCTS_H__
+#define __ROMPROPERTIES_LIBROMDATA_SATURN_STRUCTS_H__
 
 /**
  * References:
@@ -145,4 +144,4 @@ typedef enum {
 }
 #endif
 
-#endif /* __ROMPROPERTIES_LIBROMDATA_DC_STRUCTS_H__ */
+#endif /* __ROMPROPERTIES_LIBROMDATA_SATURN_STRUCTS_H__ */