+++ NEWS.md
@@ -55,6 +55,9 @@
     * Show access rights, i.e. AHBPROT and DVD Video.
     * Detect Wii update partitions that have Boot2 and a single IOS, but
       not a System Menu.
+    * Added support for NASOS images. (GCML, WII5, and WII9)
+      * GCMM images are not supported yet because there doesn't seem to
+        be any documentation on how to create one.
   * NES: Added support for the internal footer present in some ROMs. This
     footer is used for, among other things, FamicomBox.
   * Nintendo3DS: Split the SMDH code into a separate parser. This should